Blueprints now loads from main.py not __init__.py
