+++ quokka/ext/blueprints.py
@@ -30,32 +30,24 @@ def load_from_folder(app):
     dir_list = os.listdir(path)
     mods = {}
     object_name = app.config.get('BLUEPRINTS_OBJECT_NAME', 'module')
+    module_file = app.config.get('BLUEPRINTS_MODULE_NAME', 'main')
+    blueprint_module = module_file + '.py'
     for fname in dir_list:
         if not os.path.exists(os.path.join(path, fname, 'DISABLED')) and  \
                 os.path.isdir(os.path.join(path, fname)) and \
-                os.path.exists(os.path.join(path, fname, '__init__.py')):
-                # TODO: CHANGE TO main.py
+                os.path.exists(os.path.join(path, fname, blueprint_module)):
 
             # register blueprint object
-            module_name = ".".join([base_module_name, fname])
+            module_root = ".".join([base_module_name, fname])
+            module_name = ".".join([module_root, module_file])
             mods[fname] = importlib.import_module(module_name)
             blueprint = getattr(mods[fname], object_name)
-
-            if blueprint.name not in app.blueprints:
-                app.logger.info("registering blueprint: %s" % blueprint.name)
-                app.register_blueprint(blueprint)
-            else:
-                blueprint.name += str(random.getrandbits(8))
-                app.register_blueprint(blueprint)
-                app.logger.warning(
-                    "CONFLICT:%s already registered, using %s",
-                    fname,
-                    blueprint.name
-                )
+            app.logger.info("registering blueprint: %s" % blueprint.name)
+            app.register_blueprint(blueprint)
 
             # register admin
             try:
-                importlib.import_module(".".join([module_name, 'admin']))
+                importlib.import_module(".".join([module_root, 'admin']))
             except ImportError as e:
                 app.logger.info(
                     "%s module does not define admin or error: %s", fname, e
